backend/requirements.txt
PyMuPDF
pdfplumber
regex
python-stdnum
ftfy
rapidfuzz
flask
flask-cors
python-dotenv
